made requested changes
- Modified name of ROW_MAPPER
- Change name of controller
- Added rating service
